Claymore: factorize request in base class
